split up services, alpha ordering check
