+++ src/store/index.js
@@ -8,6 +8,7 @@ const middleware = [thunk];
 if (__DEV__) {
   const logger = createLogger({
     collapsed: true,
+    // predicate removes action types from being logged in the console
     predicate: (getState, action) => action.type !== 'SET_REQUEST_PENDING',
   });
